feat: Support fallback to json struct tag names
This change introduces the ability for `goql` to fall-back to using
`json` struct tag field names (via an option). By supporting this, we'll
be able to use the models generated by `gqlgen` in our repositories
automatically with `goql`, essentially allowing us to get a complete
golang GraphQL client "for free" in each service that's using `gqlgen`.